Add settings, dynamically choose update rate
Closes #2: add two settings for sample rate and sample count. It is
possible to set these such that calculating rates of change is
impossible. On the other hand, it should really help moderate system
load if needed.

Closes #13: samples are divided into buckets based on the signal's
guessed update rate. The guess can change downward (signal updates
faster than 300 ticks), but never upward (as yet). The first bucket is
always ignored and the remaining buckets' largest values are used to
determine the rate of change per minute, and the ETD, of the signal.